Move the related test case to 'kubernetes_integration_test'
